+++ app/routes/blocksuitePlayground.tsx
@@ -1,0 +1,2 @@
+export { meta } from "./docTest";
+export { default } from "./docTest";